+++ Resources/views/Form/ckeditor_widget.html.php
@@ -16,6 +16,10 @@
             CKEDITOR.plugins.addExternal('<?php echo $pluginName ?>', '<?php echo $plugin['path'] ?>', '<?php echo $plugin['filename'] ?>');
         <?php endforeach; ?>
 
+        <?php foreach ($styles as $styleName => $style): ?>
+            CKEDITOR.stylesSet.add('<?php echo $styleName ?>', <?php echo json_encode($style) ?>);
+        <?php endforeach; ?>
+
         <?php foreach ($templates as $templateName => $template): ?>
             CKEDITOR.addTemplates('<?php echo $templateName ?>', <?php echo json_encode($template) ?>);
         <?php endforeach; ?>